Fix use of routing and link
