+++ src/App/routes/index.tsx
@@ -1625,12 +1625,12 @@ const informationTechnologyServices = customWrapRoute({
 });
 
 const catalogueLivelihood = customWrapRoute({
+    parent: surgeCatalogueLayout,
     path: 'livelihood',
     component: {
         render: () => import('#views/CatalogueLivelihood'),
         props: {},
     },
-    parent: surgeCatalogue,
     wrapperComponent: Auth,
     context: {
         title: 'Livelihoods and Basic Needs',
@@ -1638,27 +1638,13 @@ const catalogueLivelihood = customWrapRoute({
     },
 });
 
-const catalogueLivelihoodIndex = customWrapRoute({
-    index: true,
-    component: {
-        render: () => import('#views/CatalogueLivelihoodIndex'),
-        props: {},
-    },
-    parent: catalogueLivelihood,
-    wrapperComponent: Auth,
-    context: {
-        title: 'Livelihood Services',
-        visibility: 'anything',
-    },
-});
-
 const livelihoodServices = customWrapRoute({
-    path: 'livelihoods-and-basic-needs',
+    parent: surgeCatalogueLayout,
+    path: 'livelihood/livelihoods-and-basic-needs',
     component: {
         render: () => import('#views/LivelihoodServices'),
         props: {},
     },
-    parent: catalogueLivelihood,
     wrapperComponent: Auth,
     context: {
         title: 'Livelihood Service',
@@ -2039,7 +2025,6 @@ const wrappedRoutes = {
     catalogueInformationTechnology,
     informationTechnologyServices,
     catalogueLivelihood,
-    catalogueLivelihoodIndex,
     threeWProjectDetail,
     livelihoodServices,
 };
